+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.1.0] - 2025-08-14
+
+### Added
+- Cat simulation feature with random cat behaviors
+- Cat method that returns random actions (meow, purr, stretch, nap, play, hunt, groom)
+- Demo implementation showing cat feature usage
+
 ## [1.0.0] - 2025-08-14
 
 ### Added